fix: id changed to int

+++ src/main/java/entities/Client.java
@@ -2,24 +2,24 @@ package entities;
 
 public class Client {
 
-	private Long id;
+	private Integer id;
 	private String name;
 	private String email;
 
 	public Client() {
 	}
 
-	public Client(Long id, String name, String email) {
+	public Client(Integer id, String name, String email) {
 		this.id = id;
 		this.name = name;
 		this.email = email;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
